requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-heroku==0.3.1
django-smart-selects==1.5.6
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
sentry-sdk==0.14.4
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
